# fixes CC-757: replace sleep 5s with sync for optimized service volume initialization